Bug fixes, optimization and code cleans

+++ Snake.cs
@@ -11,8 +11,8 @@ namespace GorbInuch
         public static char head { get; set; }
         public static char body { get; set; }
         public static int blength { get; set; }
-        public static int headx { get; set; }
-        public static int heady { get; set; }
+        public static short headx { get; set; }
+        public static short heady { get; set; }
         public sbyte Vectx { get; set; }
         public sbyte Vecty { get; set; }
 
@@ -21,7 +21,7 @@ namespace GorbInuch
             this.Vectx = vectx;
             this.Vecty = vecty;
         }
-        public static void Settings(int headx, int heady,int blength = 3,char head = '%',char body = '*')
+        public static void Settings(short headx, short heady,int blength = 3,char head = '%',char body = '*')
         {
             Snake.head = head;
             Snake.body = body;
